working graphs

+++ VS2019/ConsoleApp1/WebApplication1/Models/CoordinateData.cs
@@ -8,6 +8,6 @@ namespace WebApplication1.Models
     public class CoordinateData
     {
         public string X { get; set; }
-        public int Y { get; set; }
+        public decimal Y { get; set; }
     }
 }
